client: EmbeddingModel from_pretrained torch_dtype= -> dtype= (5.12 deprecation)
transformers 5.12.1 deprecates the torch_dtype= kwarg ("Use dtype instead!");
EmbeddingModel.__init__ still passed torch_dtype=, emitting the warning on every
embedding load (e.g. nomic-embed-text on the mobile worker). Rename to dtype= —
supported on the pinned 5.12.1 and future-proof. No behavior change. (android
client + mxfp4_convert's config field left as-is — the latter is a genuine
config.json key, not a kwarg.)

+++ client.py
@@ -2551,14 +2551,14 @@ class EmbeddingModel:
         # nomic's custom config may reject _attn_implementation="eager"; retry without it.
         try:
             self.model = AutoModel.from_pretrained(
-                model_dir, trust_remote_code=True, torch_dtype=dtype,
+                model_dir, trust_remote_code=True, dtype=dtype,
                 _attn_implementation="eager").eval()
         except ImportError:
             raise   # a MISSING dep (e.g. einops) -> let _build_with_autodeps install + retry the
             #         whole build (don't waste a 2nd no-eager attempt that hits the same ImportError)
         except Exception:
             self.model = AutoModel.from_pretrained(
-                model_dir, trust_remote_code=True, torch_dtype=dtype).eval()
+                model_dir, trust_remote_code=True, dtype=dtype).eval()
         try:
             self.model.to(device)
         except Exception:
